gcp/requirements.txt
google-api-python-client==1.8.3
python-dotenv==0.13.0
